Simplifying integration-test/tests
