+++ integration-tests/tests/src/setup-globals.ts
@@ -33,8 +33,6 @@ import { testSkipIf, suiteSkipIf } from "./utils/skip-if";
 global.describe.skipIf = suiteSkipIf;
 global.it.skipIf = testSkipIf;
 
-console.log("Patched the global");
-
 import chai from "chai";
 
 import chaiAsPromised from "chai-as-promised";
@@ -42,3 +40,25 @@ chai.use(chaiAsPromised);
 
 import { chaiRealmObjects } from "./utils/chai-plugin";
 chai.use(chaiRealmObjects);
+
+/**
+ * Use the `longTimeout` context variable to override this.
+ */
+const DEFAULT_LONG_TIMEOUT = 30 * 1000; // 30s
+
+describe("Test Harness", function (this: Mocha.Suite) {
+  /**
+   * @see [typings.d.ts](./typings.d.ts) for documentation.
+   */
+  function longTimeout(this: Mocha.Context | Mocha.Suite) {
+    this.timeout(environment.longTimeout || DEFAULT_LONG_TIMEOUT); // 30 seconds
+  }
+
+  // Patching the Suite and Context with a longTimeout method
+  // We cannot simply `import { Suite, Context } from "mocha"` here,
+  // since Mocha Remote client brings its own classes
+  const Suite = this.constructor as typeof Mocha.Suite;
+  const Context = this.ctx.constructor as typeof Mocha.Context;
+  Suite.prototype.longTimeout = longTimeout;
+  Context.prototype.longTimeout = longTimeout;
+});
